backend/pyproject.toml
[project]
name = "alphadesk-backend"
version = "0.1.0"
description = "AlphaDesk — Multi-Agent Autonomous Paper-Trading Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "alpaca-py>=0.28.0",
    "anthropic>=0.40.0",
    "pinecone>=6.0.0",
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "lxml>=5.0.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
